#32 refactor: index.ts 파일 추가
