+++ src/shared/ui/Input/Email/EmailInput.stories.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { Meta, StoryFn } from '@storybook/react';
 import { useForm } from 'react-hook-form';
 import { FormValues } from '@/shared/types/input';
-import EmailInput from './Email';
+import EmailInput from './EmailInput';
 
 export default {
   title: 'Input/EmailInput',
