+++ src/shared/ui/Input/Email/index.ts
@@ -1,0 +1,1 @@
+export { default as EmailInput } from './EmailInput';
