+++ src/shared/ui/Input/Input.stories.tsx
@@ -1,5 +1,5 @@
 import { StoryFn, Meta } from '@storybook/react';
-import Input from '.';
+import Input from './Input';
 import '@/styles/globals.css';
 
 export default {
