+++ src/shared/ui/Input/Nickname/index.ts
@@ -1,0 +1,1 @@
+export { default as NicknameInput } from './NicknameInput';
