+++ src/shared/ui/Input/Password/index.ts
@@ -1,0 +1,1 @@
+export { default as PasswordInput } from './PasswordInput';
