+++ src/shared/ui/Input/PasswordCheck/index.ts
@@ -1,0 +1,1 @@
+export { default as PasswordCheckInput } from './PasswordCheckInput';
