+++ src/shared/ui/Input/TextBox/index.ts
@@ -1,0 +1,1 @@
+export { default as TextBoxInput } from './TextBoxInput';
