+++ src/shared/ui/Input/image/index.ts
@@ -1,0 +1,1 @@
+export { default as ImageInput } from './ImageInput';
